requirements.txt
python-docx
mistune
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
cairosvg
Pillow
